* refactor Animation::ExchangeViewAnimation to better test it * add a bunch of unit tests (coverage now : 6.4% lines) * fix view container listener viewRemoved not called * write script for better unit test build workflow

+++ vstgui/lib/animation/timingfunctions.cpp
@@ -115,8 +115,8 @@ float InterpolationTimingFunction::getPosition (uint32_t milliseconds)
 			return pos;
 		else if (time <= milliseconds && prevTime > milliseconds)
 		{
-			float timePos = (float)(milliseconds - time) / (float)(prevTime - time);
-			return pos + ((prevPos - pos) * timePos);
+			double timePos = (double)(milliseconds - time) / (double)(prevTime - time);
+			return static_cast<float> (static_cast<double> (pos) + ((static_cast<double> (prevPos) - static_cast<double> (pos)) * timePos));
 		}
 		prevTime = time;
 		prevPos = pos;
